move mayor stuff to what i have to assume is the mayor api
